Move Modal-columns file to modal parts folder
